legacy_codes/legacy_diffuser/requirements.txt
diffusers>=0.21.4
numpy
opencv-python>=4.7.0.72
pillow>=9.5.0
safetensors>=0.3.1
torch
torchvision
tqdm
transformers>=4.33.3
xformers